프로필 업데이트 및 비밀번호 업데이트

+++ src/main/java/Prochat_project/model/alarm/AlarmArgs.java
@@ -9,6 +9,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class AlarmArgs {
     // user who occur alarm
-    private Integer fromUserId;
-    private Integer targetId;
+    private Long fromUserId;
+    private Long targetId;
 }
